apps/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.7.0
passlib==1.7.4
PyJWT==2.8.0
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.9
